Update Zobrist to only track p1 walls placed

+++ quoridor/src/constants.rs
@@ -1,7 +1,7 @@
 pub const MAX_NUMBER_OF_MOVES: usize = 200;
 pub const BOARD_WIDTH: usize = 9;
 pub const BOARD_HEIGHT: usize = 9;
-pub const NUM_WALLS_PER_PLAYER: usize = 10;
+pub const NUM_WALLS_PER_PLAYER: u8 = 10;
 
 pub const INPUT_H: usize = BOARD_HEIGHT;
 pub const INPUT_W: usize = BOARD_WIDTH;
